[rename] module from compose to calendar
