+++ app/src/main/java/com/boiled/calendar/sample/month/SampleMonthCalendarA.kt
@@ -21,10 +21,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.boiled.calendar.compose.component.header.WeekdaysHeader
-import com.boiled.calendar.compose.monthcalendar.MonthCalendar
-import com.boiled.calendar.compose.monthcalendar.MonthCalendarState
-import com.boiled.calendar.compose.monthcalendar.rememberMonthCalendarState
+import com.boiled.calendar.calendar.component.header.WeekdaysHeader
+import com.boiled.calendar.calendar.monthcalendar.MonthCalendar
+import com.boiled.calendar.calendar.monthcalendar.MonthCalendarState
+import com.boiled.calendar.calendar.monthcalendar.rememberMonthCalendarState
 import kotlinx.coroutines.launch
 import java.util.Locale
