+++ calendar/proguard-rules.pro
