+++ calendar/src/main/AndroidManifest.xml
