+++ calendar/src/test/java/com/boiled/calendar/calendar/ExampleUnitTest.kt
@@ -1,0 +1,17 @@
+package com.boiled.calendar.calendar
+
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
